Renaming and reorganizing package to better catkinify/pep8ify
