+++ .gitignore
@@ -1,3 +1,6 @@
 .DS_Store
 *.pyc
 .idea
+*~
+build
+devel
