+++ src/ros_tcp_endpoint/communication.py
@@ -1,0 +1,35 @@
+#  Copyright 2020 Unity Technologies
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+
+class RosSender:
+    """
+        Base class for ROS communication where data is sent to the ROS network.
+    """
+    def __init__(self):
+        pass
+
+    def send(self, *args):
+        raise NotImplementedError
+
+
+class RosReceiver:
+    """
+        Base class for ROS communication where data is being sent outside of the ROS network.
+    """
+    def __init__(self):
+        pass
+
+    def send(self, *args):
+        raise NotImplementedError
