+++ src/ros_tcp_endpoint/publisher.py
@@ -1,15 +1,27 @@
-#!/usr/bin/env python
+#  Copyright 2020 Unity Technologies
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 
 import rospy
 
-from tcp_endpoint.RosCommunication import RosSender
+from .communication import RosSender
 
 
 class RosPublisher(RosSender):
     """
     Class to publish messages to a ROS topic
     """
-    # TODO: surrface latch functionality
+    # TODO: surface latch functionality
     def __init__(self, topic, message_class, queue_size=10):
         """
 
@@ -18,6 +30,7 @@ class RosPublisher(RosSender):
             message_class: The message class in catkin workspace
             queue_size:    Max number of entries to maintain in an outgoing queue
         """
+        RosSender.__init__(self)
         self.msg = message_class()
         self.pub = rospy.Publisher(topic, message_class, queue_size=queue_size)
