+++ src/ros_tcp_endpoint/server.py
@@ -1,4 +1,16 @@
-#!/usr/bin/env python
+#  Copyright 2020 Unity Technologies
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 
 import rospy
 import socket
@@ -6,6 +18,8 @@ import socket
 from tcp_endpoint.UnityTCPSender import UnityTCPSender
 from tcp_endpoint.RosTCPClientThread import ClientThread
 from tcp_endpoint.RosUnityHandshakeService import RosUnityHandshakeService
+from .client import ClientThread
+
 
 class TCPServer:
     """
@@ -57,9 +71,10 @@ class TCPServer:
             new_thread = ClientThread(conn, self, ip, port)
             new_thread.start()
             threads.append(new_thread)
-
-        for t in threads:
-            t.join()
+        
+        # Unreachable statements:
+        # for t in threads:
+        #    t.join()
 
     def send_unity_error(self, error):
         self.unity_tcp_sender.send_unity_error(error)
