+++ src/ros_tcp_endpoint/subscriber.py
@@ -1,10 +1,22 @@
-#!/usr/bin/env python
+#  Copyright 2020 Unity Technologies
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 
 import rospy
 import socket
 
-from tcp_endpoint.RosCommunication import RosReceiver
-from tcp_endpoint.RosTCPClientThread import ClientThread
+from .communication import RosReceiver
+from .client import ClientThread
 
 
 class RosSubscriber(RosReceiver):
@@ -20,6 +32,7 @@ class RosSubscriber(RosReceiver):
             message_class: The message class in catkin workspace
             queue_size:    Max number of entries to maintain in an outgoing queue
         """
+        RosReceiver.__init__(self)
         self.topic = topic
         self.node_name = "{}_subsciber".format(topic)
         self.msg = message_class
