--- tcp_endpoint/src/setup.py
@@ -1,11 +1,0 @@
-## ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD
-
-from setuptools import setup
-from catkin_pkg.python_setup import generate_distutils_setup
-
-# fetch values from package.xml
-setup_args = generate_distutils_setup(
-    packages=['tcp_endpoint'],
-    package_dir={'': 'src'})
-
-setup(**setup_args)
