--- tcp_endpoint/src/tcp_endpoint/RosCommunication.py
@@ -1,23 +1,0 @@
-#!/usr/bin/env python
-
-
-class RosSender:
-    """
-        Base class for ROS communication where data is sent to the ROS network.
-    """
-    def __init__(self):
-        pass
-
-    def send(self):
-        pass
-
-
-class RosReceiver:
-    """
-        Base class for ROS communication where data is being sent outside of the ROS network.
-    """
-    def __init__(self):
-        pass
-
-    def send(self):
-        pass
